+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versions follow [Semantic Versioning](https://semver.org/spec/v2.0.0.html) (`<ma
 * Bump vendored libraries to v10.2 compatible `*.mpy` files
 * Bump vendored `mpy-cross` binaries to CircuitPython v10.2
 * #47 Update OpenSky API for new authentication flow
+* Migrate from `settings.py` to `settings.toml` for secrets management
 
 ## [v2.0.0]
 ### Added